backend/pyproject.toml
[project]
name = "mica-backend"
version = "0.4.0"
description = "Mica | 觅采 — Internal procurement management system backend"
requires-python = ">=3.12"
license = "Apache-2.0"
authors = [
    { name = "Mica Contributors" }
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "gunicorn>=23.0.0",
    "sqlalchemy[asyncio]>=2.0.35",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "email-validator>=2.2.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.17",
    "cryptography>=43.0.0",
    "python-dateutil>=2.9.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.2",
    "litellm>=1.52.0",
    "aiofiles>=24.1.0",
    "pdfplumber>=0.11.4",
    "pypdf>=5.1.0",
    "pymupdf>=1.25.1",
    "easyofd>=0.5.6",
    "lxml>=5.3.0",
    "reportlab>=4.2.5",
    "openpyxl>=3.1.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "RUF", # ruff-specific rules
]
ignore = [
    "E501", # line too long (handled by formatter)
    "B008", # function call in argument defaults (FastAPI uses this)
    "RUF001", # ambiguous unicode (CJK prompts + i18n strings legitimately use fullwidth punctuation)
    "RUF002", # ambiguous unicode in docstrings (same reason)
    "RUF003", # ambiguous unicode in comments (same reason)
]

[tool.ruff.lint.per-file-ignores]
"migrations/**" = ["E", "F", "I"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
strict = false  # Relaxed during Walking Skeleton phase
warn_redundant_casts = true
warn_unused_ignores = true
disallow_untyped_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_return_any = false
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["migrations.*", "tests.*"]
ignore_errors = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "integration: integration tests requiring database",
    "unit: pure unit tests",
]
